LayoutTests/fast/forms/input-step-expected.txt

Test the behavior of step attribute and step DOM property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.step is ""
PASS input.getAttribute("step") is "foo"
PASS input.step is "bar"
PASS input.step is ""
PASS input.getAttribute("step") is null
PASS input.step is "null"
PASS input.step is "undefined"
PASS input.getAttribute("step") is "undefined"
PASS input.step is "undefined"
PASS input.step is "256"
PASS input.getAttribute("step") is "256"
PASS input.step is "256"
PASS successfullyParsed is true

TEST COMPLETE
